+++ src/Public/Get-Accelerator.ps1
@@ -1,3 +1,19 @@
 ﻿function Get-Accelerator {
-    [psobject].Assembly.GetType('System.Management.Automation.TypeAccelerators')::Get
+    # .EXTERNALHELP PWAddins-help.xml
+    [CmdletBinding()]
+    param (
+            [SupportsWildcards()]
+            [string]
+        $Name = '*'
+    )
+
+    $TAType = [psobject].Assembly.GetType('System.Management.Automation.TypeAccelerators')
+    $AcceleratorList = $TAType::Get
+    $AcceleratorList.Keys | Where-Object { $_ -like $Name } | ForEach-Object {
+        $PropList = [ordered] @{
+            Name = $_
+            Type = if ($_ -like 'psobject') { [psobject] } else { $AcceleratorList.$_ }
+        }
+        New-Object -TypeName PSObject -Property $PropList
+    }
 }